pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adele-judge"
version = "0.1.0"
description = "Configurable ADeLe-specific distilled judge training and evaluation pipeline."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "datasets>=2.19",
  "numpy>=1.26",
  "pandas>=2.1",
  "pyarrow>=15",
  "pyyaml>=6",
  "scikit-learn>=1.4",
  "tqdm>=4.66",
  "transformers>=4.51",
]

[project.optional-dependencies]
train = [
  "accelerate>=0.34",
  "bitsandbytes>=0.43",
  "peft>=0.12",
  "torch>=2.3",
  "trl>=0.9",
  "unsloth>=2024.11",
]
dev = [
  "pytest>=8",
  "ruff>=0.6",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
